% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl_logo_badge.R
\name{dsl_badge}
\alias{dsl_badge}
\title{JHU DSL Badge}
\usage{
dsl_badge(logo = dsl_logo_url(), alt_text = "JHU Data Science")
}
\arguments{
\item{logo}{path to the logo}

\item{alt_text}{Alternate text for the badge if not found}
}
\value{
Character vector
}
\description{
JHU DSL Badge
}
\examples{
dsl_badge()
}
