% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dsl_logo.R
\name{use_dsl_logo}
\alias{use_dsl_logo}
\title{Use JHU DSL Logo}
\usage{
use_dsl_logo(base_path = ".", ...)
}
\arguments{
\item{base_path}{Path to root directory}

\item{...}{arguments to pass to \code{\link{dsl_badge}}}
}
\value{
Invisible TRUE
}
\description{
Use JHU DSL Logo
}
